engine-server/api_gateway/api/serializers.py
from rest_framework import serializers
from .models import Tweet, UserProfile

class TweetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tweet
        fields = '__all__'

class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = '__all__'
