engine-server/api_gateway/requirements.txt
Django>=4.0
djangorestframework
djongo
pymongo
celery
redis
